# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Document Center API",
    "summary": "Документ-центр ИП «Северный Контур»: проекты, документы, версии и аудит",
    "description": "\nВнутренний REST API ИП «Северный Контур» для управления документами по проектам.\n\n**Базовый путь публичного контракта:** `/api/v1`. Версия контракта (`v1`)\nне совпадает с версией приложения, указанной в заголовке документации.\n\n### Аутентификация\nПолучите токен операцией `POST /api/v1/auth/login` и передавайте его в заголовке\n`Authorization: Bearer <token>`. В Swagger UI нажмите **Authorize** и вставьте токен.\n\n### Роли\n- **admin** — полный доступ ко всем проектам, пользователям и журналу действий;\n- **manager** — создаёт проекты и выдаёт доступ участникам своих проектов;\n- **worker** — работает с документами в рамках выданного доступа `editor`;\n- **viewer** — только чтение в рамках выданного доступа.\n\n### Ошибки\nВсе ошибки возвращаются в формате RFC 9457 Problem Details\n(`application/problem+json`) с полями `type`, `title`, `status`, `detail`,\n`instance`; ошибки валидации дополнительно содержат массив `errors`.\n\n### Коллекции\nКоллекции поддерживают параметры `offset` (≥ 0) и `limit` (1…100, по умолчанию 20)\nи возвращают конверт `items`, `total`, `offset`, `limit`.\n",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/documents": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Получить документы проекта",
        "description": "Вернуть документы проекта, доступного пользователю.",
        "operationId": "list_project_documents",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Идентификатор проекта",
              "title": "Project Id"
            },
            "description": "Идентификатор проекта"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DocumentStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Отбор по состоянию документа",
              "title": "Status"
            },
            "description": "Отбор по состоянию документа"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Смещение от начала коллекции",
              "default": 0,
              "title": "Offset"
            },
            "description": "Смещение от начала коллекции"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Размер страницы, максимум 100",
              "default": 20,
              "title": "Limit"
            },
            "description": "Размер страницы, максимум 100"
          }
        ],
        "responses": {
          "200": {
            "description": "Страница документов проекта",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_DocumentRead_"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Создать документ",
        "description": "Создать документ в проекте.\n\nДокумент создаётся в состоянии draft, его содержимое сохраняется\nкак первая версия. Требуется уровень доступа editor.",
        "operationId": "create_document",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Идентификатор проекта",
              "title": "Project Id"
            },
            "description": "Идентификатор проекта"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Созданный документ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRead"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Получить документ",
        "description": "Вернуть документ, доступный текущему пользователю.\n\nДокумент недоступного проекта представляется как несуществующий.",
        "operationId": "get_document",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Идентификатор документа",
              "title": "Document Id"
            },
            "description": "Идентификатор документа"
          }
        ],
        "responses": {
          "200": {
            "description": "Карточка документа",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRead"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Document not found",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Documents"
        ],
        "summary": "Изменить документ",
        "description": "Изменить название, содержимое или состояние документа.\n\nПубликация и архивирование выполняются изменением поля ``status``.\nИзменение содержимого создаёт новую версию. Требуется уровень\nдоступа editor.",
        "operationId": "update_document",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Идентификатор документа",
              "title": "Document Id"
            },
            "description": "Идентификатор документа"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Изменённый документ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRead"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Document not found",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Invalid document state transition",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Documents"
        ],
        "summary": "Удалить документ",
        "description": "Удалить документ вместе со всеми его версиями.",
        "operationId": "delete_document",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Идентификатор документа",
              "title": "Document Id"
            },
            "description": "Идентификатор документа"
          }
        ],
        "responses": {
          "204": {
            "description": "Документ удалён"
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Document not found",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/versions": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Получить версии документа",
        "description": "Вернуть сохранённые версии документа.",
        "operationId": "list_document_versions",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Идентификатор документа",
              "title": "Document Id"
            },
            "description": "Идентификатор документа"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Смещение от начала коллекции",
              "default": 0,
              "title": "Offset"
            },
            "description": "Смещение от начала коллекции"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Размер страницы, максимум 100",
              "default": 20,
              "title": "Limit"
            },
            "description": "Размер страницы, максимум 100"
          }
        ],
        "responses": {
          "200": {
            "description": "Страница версий в порядке убывания номера",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_DocumentVersionReadWithCreator_"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Document not found",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Восстановить версию документа",
        "description": "Создать версию документа из ранее сохранённой.\n\nИстория версий не изменяется: содержимое выбранной версии становится\nтекущим и сохраняется как новая версия. Требуется уровень доступа\neditor.",
        "operationId": "restore_document_version",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Идентификатор документа",
              "title": "Document Id"
            },
            "description": "Идентификатор документа"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentVersionRestore"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Документ с восстановленным содержимым",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRead"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Document version not found",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/versions/{version}": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Получить версию документа",
        "description": "Вернуть конкретную версию документа.",
        "operationId": "get_document_version",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Идентификатор документа",
              "title": "Document Id"
            },
            "description": "Идентификатор документа"
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Порядковый номер версии",
              "title": "Version"
            },
            "description": "Порядковый номер версии"
          }
        ],
        "responses": {
          "200": {
            "description": "Сохранённая версия документа",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentVersionRead"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Document version not found",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DocumentCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 120,
            "minLength": 3,
            "title": "Title",
            "description": "Название документа, от 3 до 120 символов",
            "examples": [
              "Пояснительная записка"
            ]
          },
          "content": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100000
              },
              {
                "type": "null"
              }
            ],
            "title": "Content",
            "description": "Содержимое документа, не более 100000 символов",
            "default": "",
            "examples": [
              "Раздел 1. Общие положения"
            ]
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "DocumentCreate",
        "description": "Данные для создания документа."
      },
...
      "DocumentStatus": {
        "type": "string",
        "enum": [
          "draft",
          "published",
          "archived"
        ],
        "title": "DocumentStatus",
        "description": "Состояние жизненного цикла документа: draft, published или archived."
      },
...
      "Page_DocumentRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DocumentRead"
            },
            "type": "array",
            "title": "Items",
            "description": "Записи текущей страницы"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total",
            "description": "Общее количество записей, доступных запросу"
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset",
            "description": "Смещение текущей страницы от начала коллекции"
          },
          "limit": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "description": "Запрошенный размер страницы"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "Page[DocumentRead]",
        "description": "Страница коллекции с метаданными пагинации: items, total, offset, limit"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}